new directory structure visitor/cyclic visitor/acyclic
