+++ VisitorFramework/include/Visitor/Acyclic/TypeFunctions.h
@@ -24,9 +24,9 @@ struct EmptyAccessor{
 template<class ToVisit, class Accessor_ = EmptyAccessor>
 class VisitorInterface{
 public:
-	std::string toString(){ return "VisitableAdapter";}
+	std::string toString(){ return "VisitorInterface";}
 
-	void print(){ std::cout << "VisitableAdapter" << std::endl; }
+	void print(){ std::cout << "VisitorInterface" << std::endl; }
 	using ConcreteVisitable = ToVisit;
 	using Accessor = Accessor_;
 	using this_type = VisitorInterface<ToVisit, Accessor>;
@@ -64,18 +64,17 @@ template<class ToVisit, class =
 								>
 		>
 struct getVisitorType{
-	// kein Visitor und Accessor in ToVisit verfügbar
+	// no Visitor and no Accessor in ToVisit available
 	using type = typename VisitorInterface<ToVisit>::type;
 };
 template<class ToVisit>
 struct getVisitorType<ToVisit, typename hasVisitor<ToVisit>::type>{
-	// ToVisit definiert seinen Visitor
+	// ToVisit defines its Visitor
 	using type = typename hasVisitor<ToVisit>::type;
-
 };
 template<class ToVisit>
 struct getVisitorType<ToVisit, typename hasAccessor<ToVisit>::type>{
-	// ToVisit definiert nur einen Accessor
+	// ToVisit defines only an Accessor
 	using type = typename VisitorInterface<ToVisit, typename ToVisit::Accessor>::type;
 };
