+++ VisitorFramework/include/Visitor/Cyclic/VisitorCyclic.h
@@ -8,7 +8,9 @@
 #ifndef VISITORCYCLIC_H_
 #define VISITORCYCLIC_H_
 
-#include "StoragePolicies.h"
+#include "../StoragePolicies.h"
+#include "../BaseKind.h"
+#include "../MakeTypelist.h"
 
 #include <iostream>
 
@@ -151,6 +153,45 @@ using visitsAbstract = InheritFromAbstract<ToVisit, Rest...>;
 
 
 
+
+template<class LoggingPolicy_, class = BaseKind::Abstract>
+struct VisitorBase{
+	template<class ...Visitables>
+	using implementsVisitor = VisitorCyclic::InheritFromAbstract<Visitables...>;
+};
+template<class LoggingPolicy_>
+struct VisitorBase<LoggingPolicy_, BaseKind::Default>{
+	template<class ...Visitables>
+	using implementsVisitor = VisitorCyclic::InheritFromDefault<LoggingPolicy_, Visitables...>;
+};
+
+template<class LoggingPolicy, class BaseKind_, class ...Visitables>
+struct Repository{
+	using VisitorBase =
+			typename VisitorBase<LoggingPolicy, BaseKind_>::template
+			implementsVisitor<Visitables...>;
+
+	using Visitable = VisitorCyclic::Visitable<VisitorBase>;
+
+	template<class ConcreteVisitable>
+	using VisitableImpl =
+			VisitorCyclic::VisitableImpl<ConcreteVisitable, VisitorBase, LoggingPolicy>;
+
+	template<class Adaptee, class StoragePolicy>
+	using VisitableAdapter =
+			VisitorCyclic::VisitableAdapter<Adaptee, StoragePolicy, LoggingPolicy, VisitorBase>;
+
+	// Convenience Interface
+	template<class Adaptee>
+	using AdapterByWeakpointer = VisitableAdapter<Adaptee, StorageByWeakPointer<Adaptee>>;
+	template<class Adaptee>
+	using AdapterByReference = VisitableAdapter<Adaptee, StorageByReference<Adaptee>>;
+};
+
+template<class LoggingPolicy, class BaseKind_, class ...Visitables>
+struct Repository<LoggingPolicy, BaseKind_, Visitor::MakeTypelist<Visitables...>>
+// delegates
+: Repository<LoggingPolicy, BaseKind_, Visitables...>{};
